engine: split implementation from header
