+++ CMakeLists.txt
@@ -1,13 +1,17 @@
 cmake_minimum_required(VERSION 3.20)
 project(tank-battle)
 
+add_library(engine "" "")
+target_include_directories(engine PUBLIC engine)
+target_sources(engine PRIVATE engine/engine.c)
+
 add_executable(tankbattle "" "")
 target_include_directories(tankbattle PRIVATE tankbattle)
 target_sources(tankbattle PRIVATE
 	tankbattle/tankbattle.c
         tankbattle/calc.c
 )
-target_link_libraries(tankbattle SDL2 m)
+target_link_libraries(tankbattle SDL2 m engine)
 
 # tests
 add_subdirectory(unity)
@@ -17,8 +21,7 @@ target_include_directories(test-calc PRIVATE tankbattle)
 target_link_libraries(test-calc PRIVATE unity m)
 
 add_executable(test-engine tests/engine.c)
-target_include_directories(test-engine PRIVATE engine)
-target_link_libraries(test-engine PRIVATE unity SDL2)
+target_link_libraries(test-engine PRIVATE unity SDL2 engine)
 
 include(CTest)
 add_test(NAME calc COMMAND test-calc)
